solana_clap_utils

Module keypair

Loading signers and keypairs from the command line.

This module contains utilities for loading Signers and Keypairs from standard signing sources, from the command line, as in the Solana CLI.

The key function here is signer_from_path, which loads a Signer from one of several possible sources by interpreting a “path” command line argument. Its documentation includes a description of all possible signing sources supported by the Solana CLI. Many other functions here are variations on, or delegate to, signer_from_path.

Structs

CliSignerInfo

DefaultSigner

A command line argument that loads a default signer in absence of other signers.

SignOnly

SignerFromPathConfig

Constants

ASK_KEYWORD

SKIP_SEED_PHRASE_VALIDATION_ARG

Functions

keypair_from_path

Loads a Keypair from one of several possible sources.

keypair_from_seed_phrase

Reads user input from stdin to retrieve a seed phrase and passphrase for keypair derivation.

presigner_from_pubkey_sigs

prompt_passphrase

Prompts user for a passphrase and then asks for confirmirmation to check for mistakes

pubkey_from_path

Loads the pubkey of a Signer from one of several possible sources.

resolve_signer_from_path

signer_from_path

Loads a Signer from one of several possible sources.

signer_from_path_with_config

Loads a Signer from one of several possible sources.

Type Aliases

CliSigners

SignerIndex